后端重构完成

+++ blog-api/src/main/java/cn/isbut/controller/admin/LoginLogController.java
@@ -54,7 +54,7 @@ public class LoginLogController {
 	 * @return result
 	 */
 	@DeleteMapping("/loginLog")
-	public Result delete(@RequestParam Long id) {
+	public Result delete(@RequestParam Integer id) {
 		loginLogService.deleteLoginLogById(id);
 		return Result.ok("删除成功");
 	}
